fixing circular dependency
